+++ noodles-tabix/examples/tabix_list_reference_sequence_names_async.rs
@@ -14,7 +14,11 @@ async fn main() -> io::Result<()> {
     let tabix_src = format!("{src}.tbi");
     let index = tabix::r#async::read(tabix_src).await?;
 
-    for reference_sequence_name in index.header().reference_sequence_names() {
+    let header = index
+        .header()
+        .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidData, "missing tabix header"))?;
+
+    for reference_sequence_name in header.reference_sequence_names() {
         println!("{reference_sequence_name}");
     }