+++ noodles-tabix/src/async.rs
@@ -48,8 +48,13 @@ where
 /// #
 /// # #[tokio::main]
 /// # async fn main() -> io::Result<()> {
+/// use noodles_csi as csi;
 /// use noodles_tabix as tabix;;
-/// let index = tabix::Index::default();
+///
+/// let index = csi::Index::builder()
+///     .set_header(csi::index::Header::default())
+///     .build();
+///
 /// tabix::r#async::write("sample.vcf.gz.tbi", &index).await?;
 /// # Ok(())
 /// # }