+++ noodles-tabix/src/lib.rs
@@ -32,12 +32,14 @@ pub mod index;
 mod reader;
 mod writer;
 
-pub use self::{index::Index, reader::Reader, writer::Writer};
+pub use self::{reader::Reader, writer::Writer};
 
 #[cfg(feature = "async")]
 pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};
 
 use std::{fs::File, io, path::Path};
+
+use noodles_csi::Index;
 
 static MAGIC_NUMBER: &[u8] = b"TBI\x01";
 
@@ -71,8 +73,13 @@ where
 ///
 /// ```no_run
 /// # use std::io;
+/// use noodles_csi as csi;
 /// use noodles_tabix as tabix;
-/// let index = tabix::Index::default();
+///
+/// let index = csi::Index::builder()
+///     .set_header(csi::index::Header::default())
+///     .build();
+///
 /// tabix::write("sample.vcf.gz.tbi", &index)?;
 /// # Ok::<(), io::Error>(())
 /// ```